pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "mqtt-scale-garmin-bridge"
version = "0.1.0"
description = "Bridge between Xiaomi Mi Body Composition Scale (via MQTT/OpenMQTTGateway) and Garmin Connect"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.10"
authors = [
    {name = "Jacek", email = "jacek@example.com"}
]
keywords = ["mqtt", "garmin", "xiaomi", "scale", "body-composition", "health"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: End Users/Desktop",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Home Automation",
]

dependencies = [
    "paho-mqtt>=2.0.0",
    "garminconnect>=0.2.19",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "ruff>=0.1.0",
    "commitizen>=4.10.0",
    "black>=25.0.0",
    "flake8>=7.0.0",
    "mypy>=1.19.0",
    "isort>=7.0.0",
    "bandit>=1.9.0",
    "pre-commit>=4.5.0",
    "safety>=3.7.0",
]

[project.scripts]
mqttscale-garmin = "src.main:main"

[project.urls]
Homepage = "https://github.com/jacek/mqtt-scale-garmin-bridge"

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]

[tool.ruff]
line-length = 200
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "W"]
ignore = ["E501"]

# =============================================================================
# COMMITIZEN CONFIGURATION
# =============================================================================
[tool.commitizen]
name = "cz_conventional_commits"
version = "0.1.0"
tag_format = "v$version"
version_files = [
    "pyproject.toml:version"
]
changelog_file = "CHANGELOG.md"
changelog_start_rev = "v0.1.0"
style = [
    ["qmark", "fg:#ff9d00 bold"],
    ["question", "bold"],
    ["answer", "fg:#ff9d00 bold"],
    ["pointer", "fg:#ff9d00 bold"],
    ["highlighted", "fg:#ff9d00 bold"],
    ["selected", "fg:#cc5454"],
    ["separator", "fg:#cc5454"],
    ["instruction", ""],
    ["text", ""],
    ["disabled", "fg:#858585 italic"]
]

# =============================================================================
# ISORT CONFIGURATION
# =============================================================================
[tool.isort]
profile = "black"
multi_line_output = 3
line_length = 100
known_first_party = ["src"]
skip_glob = ["**/.venv/**", "**/venv/**"]

# =============================================================================
# MYPY CONFIGURATION
# =============================================================================
[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false  # Start with false, can be strict later
disallow_incomplete_defs = false
check_untyped_defs = true
disallow_untyped_decorators = false
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true
strict_equality = true

[[tool.mypy.overrides]]
module = [
    "paho.mqtt.*",
    "garminconnect.*",
    "garth.*",
    "yaml.*"
]
ignore_missing_imports = true

# =============================================================================
# BANDIT CONFIGURATION
# =============================================================================
[tool.bandit]
exclude_dirs = ["tests", ".venv", "venv"]
skips = ["B101", "B601"]  # Skip assert and shell usage tests
